backend/requirements.txt
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-Cors==6.0.2
python-dotenv==1.2.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
google-auth==2.47.0
google-auth-oauthlib==1.2.4
requests==2.32.5
openai==2.15.0
deepl==1.27.0
deep-translator==1.11.4
PyPDF2==3.0.1
PyMuPDF>=1.24.0
pdf2docx>=0.5.8
python-docx==1.2.0
openpyxl==3.1.5
fpdf==1.7.2
beautifulsoup4==4.12.2
Werkzeug==3.1.5
pytesseract==0.3.10
Pillow==10.4.0
opencv-python-headless>=4.10.0.0
numpy>=2.1.0
